refactor: rename to ComponentSomethingButton
