+++ frontend/svelte/src/lib/components/accounts/HardwareWalletShowActionButton.svelte
