+++ frontend/svelte/src/lib/components/accounts/RenameSubAccountButton.svelte
